+++ R/get_varcov.R
@@ -338,6 +338,19 @@ get_varcov.cpglm <- get_varcov.cpglmm
 
 
 
+#' @export
+get_varcov.cglm <- function(x, ...) {
+  vc <- as.matrix(x$var)
+
+  if (.is_negativ_matrix(vc)) {
+    vc <- .fix_negative_matrix(vc)
+  }
+
+  .remove_backticks_from_matrix_names(as.matrix(vc))
+}
+
+
+
 #' @rdname get_varcov
 #' @export
 get_varcov.mixor <- function(x, effects = c("all", "fixed", "random"), ...) {